fixed buggy backprop by replacing float32 w float16 in the augmentations + other bugfixes

+++ utils/transforms.py
@@ -26,8 +26,9 @@ to_tensor = v2.Compose([v2.ToImage(), v2.ToDtype(torch.float32, scale=True)])
 
 # Define standard transformations
 img_std_transformations = {
-    "std_cityscapes" : v2.Compose([v2.Resize((512, 1024), Image.BILINEAR),v2.ToImage(), v2.ToDtype(torch.float32, scale=True),v2.Normalize(mean=MEAN, std=STD)]),
-    "std_gta5" : v2.Compose([v2.Resize((526,957), Image.BILINEAR),v2.ToImage(), v2.ToDtype(torch.float32, scale=True),v2.Normalize(mean=MEAN, std=STD)])
+    # this can be source of error float16 should be the datatype (before it was float32)
+    "std_cityscapes" : v2.Compose([v2.Resize((512, 1024), Image.BILINEAR),v2.ToImage(), v2.ToDtype(torch.float16, scale=True),v2.Normalize(mean=MEAN, std=STD)]),
+    "std_gta5" : v2.Compose([v2.Resize((526,957), Image.BILINEAR),v2.ToImage(), v2.ToDtype(torch.float16, scale=True),v2.Normalize(mean=MEAN, std=STD)])
 }
 
 lbl_std_transformations = {
